# News/api.yml
openapi: 3.0.0
info:
  title: HackerNews API
  description: API for HackerNews clone with user profiles, submissions, and comments
  termsOfService: https://www.google.com/policies/terms/
  contact:
    email: contact@news.local
  license:
    name: MIT
  version: v1
servers:
  - url: http://127.0.0.1:8000/api
    description: Local development server
  - url: https://hackernews-jwl9.onrender.com/api
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Api-Key
      in: header
  schemas:
...
    SubmissionList:
      type: object
      required:
        - title
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 255
          minLength: 1
        url:
          type: string
          format: uri
          maxLength: 500
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        points:
          type: integer
        user:
          type: string
          readOnly: true
...
security:
  - ApiKeyAuth: []
paths:
...
  /submissions/search/:
    get:
      operationId: search_submissions
      description: Search for submissions by title
      parameters:
        - name: Title of the Submission
          in: query
          description: Search query for submission titles
          required: false
          schema:
            type: string
      responses:
        '200':
          description: A list of submissions matching the search query
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubmissionList'
        '404':
          description: No submissions found
      tags:
        - submissions